fiz correcoes

+++ api/video.py
@@ -1,8 +1,6 @@
 import os, sys, traceback, re, cv2, time, uuid, shutil, traceback, hashlib, json, base64, requests;
-import cv2, sys, os, re, subprocess, uuid;
+import subprocess, threading, argostranslate, torch;
 import soundfile as sf
-import torch
-import threading, argostranslate
 
 #           VOICES              
 #https://huggingface.co/hexgrad/Kokoro-82M/blob/main/VOICES.md#italian
@@ -13,12 +11,14 @@ from kokoro import KPipeline
 from IPython.display import display, Audio
 #from googletrans import Translator
 #https://gtts.readthedocs.io/en/latest/module.html#localized-accents
-from gtts import gTTS
-from moviepy.editor import *
+from gtts import gTTS  # LIXO QUE VOU RREMOVER, ESSA BOSTA NÁO É OFF
+from moviepy.editor import *    # É UMA BOSTA CHEIA DE MEMORY LEAK
 from faster_whisper import WhisperModel
 from datetime import datetime as dt
 
 #https://www.argosopentech.com/argospm/index/
+# 1 - Palavras reservadas  <<< dicionario de palavras
+# 2 - não leva em consideraçào o contexto.... << retreinar o argos tranlator
 from argostranslate import package, translate
 # pip3 install IPython
 # pip3 install kokoro
@@ -40,8 +40,6 @@ CONFIG_GOOGLE = None;
 if os.path.exists(os.path.expandvars("$HOME/.google.json")):
     CONFIG_GOOGLE = json.loads(open(os.path.expandvars("$HOME/.google.json"), "r").read());
 
-
-
 class Video:
     def __init__(self, project, path_video):
         self.project = project;
@@ -49,11 +47,8 @@ class Video:
         self.path_video = path_video;
         self.filename   = self.path_video[self.path_video.rfind("/") + 1: self.path_video.rfind(".")];
         self.directory  = os.path.dirname( self.path_video );
-        #self.path_dir_saida = self.path_video[:self.path_video.rfind("/")] + "/ptbr/";
         self.legendas = [];
-        #self.path_final_video = os.path.join( self.path_dir_saida, self.filename + ".mkv" );
         self.path_final_vtt =   os.path.join( self.directory,  self.filename + ".vtt" );
-        #os.makedirs(self.path_dir_saida, exist_ok=True);
         self.tmp_path = "/tmp/file_" + self.id + ".kvm"
         cam = cv2.VideoCapture( path_video );
         self.fps = cam.get(cv2.CAP_PROP_FPS);
@@ -70,15 +65,13 @@ class Video:
             path_language = os.path.join( self.directory, language["directory"]);
             if not os.path.exists(path_language):
                 os.makedirs( path_language , exist_ok=True);  
-            #if self.existe():
-            #    self.started = False;
+
     def existe(self, language_dir = None):
         if language_dir == None:
             for language in self.project.languages:
                 dir_final = os.path.join( self.directory, language["directory"] );
                 file_mkv = os.path.join( dir_final, self.filename + ".mkv");
                 file_mp4 = os.path.join( dir_final, self.filename + ".mp4");
-                print(os.path.exists(file_mp4), file_mp4);
                 if not os.path.exists(file_mp4):
                     return False;
             return True;
@@ -86,17 +79,6 @@ class Video:
             dir_final = os.path.join( self.directory, language_dir );
             file_mp4 = os.path.join( dir_final, self.filename + ".mp4");
             return os.path.exists(file_mp4);
-        #for language in self.project.languages:
-        #    dir_final = os.path.join( self.directory, language["directory"] );
-        #    if not os.path.exists(dir_final):
-        #        return False;
-        #    file = os.path.join( dir_final, self.filename + ".mkv");
-        #    if os.path.exists(file):
-        #        return True;
-        #    file = os.path.join( dir_final, self.filename + ".mp4");
-        #    if os.path.exists(file):
-        #        return True;
-        #return False;
 
     def legendar(self):
         if os.path.exists( self.path_final_vtt ):
@@ -108,12 +90,7 @@ class Video:
 
     def __translate__(self, legenda, to_language):
         try:
-            #if self.project.language == to_language["language"]:
-            #    legenda.traducoes[self.project.language] = legenda.texto;
-            #    print(legenda.texto, end=', ');
-            #    return legenda.texto;
             retorno = legenda.translate( self.project.language, to_language );
-            print(retorno, end=', ');
         except:
             traceback.print_exc();
         finally:
@@ -143,17 +120,16 @@ class Video:
                     with open(path_legenda, "w") as f:
                         for legenda in self.legendas:
                             f.write( legenda.traducoes[language["language"]] );
-
         except:
             traceback.print_exc();
 
-    def __make_video__(self, legenda, to_language ):
-        try:
-            retorno = legenda.make_video( self.path_video, to_language, self.project.effects ); 
-        except:
-            traceback.print_exc();
-        finally:
-            self.THREADS_TTS_CONTADOR = self.THREADS_TTS_CONTADOR - 1;
+    #def __make_video__(self, legenda, to_language ):
+    #    try:
+    #        retorno = legenda.make_video( self.path_video, to_language, self.project.effects ); 
+    #    except:
+    #        traceback.print_exc();
+    #    finally:
+    #        self.THREADS_TTS_CONTADOR = self.THREADS_TTS_CONTADOR - 1;
 
     def make_video(self):
         try:
@@ -167,34 +143,24 @@ class Video:
                 file_final = os.path.join( self.directory, language["directory"], self.filename + ".mp4" );
                 if os.path.exists(file_final):
                     continue;
-                #for legenda in self.legendas:
-                for i in range(len(self.legendas)):
-                    if i < len(self.legendas) - 1:
-                        self.legendas[i].fim_milisegundo = self.legendas[i + 1].inicio_milisegundo;
-                    #objeto = threading.Thread(target=self.__make_video__, args=(self.legendas[i], language,));
-                    #objeto.start();
-                    #semaforo.append(objeto);
-                    #if len(semaforo) >= self.project.max_threads:
-                    #    for item in semaforo:
-                    #        item.join();
-                    #    for i in range(len(semaforo)):
-                    #        semaforo[i] = None;
-                    #    semaforo = [];   
-                    self.__make_video__(self.legendas[i], language);
-                    print(str(len(self.legendas) - 1), "/", i);
-                # se for multi thread tem que esperar as threads aqui....
-                #for item in semaforo:
-                #    item.join();
                 with open("/tmp/"+ self.id +".txt", "w") as f:
-                    contador_slice_video = 0; # para contar quantos peda´cos ja foram adicionados para que eu possa
-                    # posicionar a propaganda no video
-                    for legenda in self.legendas:
-                        contador_slice_video = contador_slice_video + 1;
-                        #if contador_slice_video == 6:
-                        #    f.write("file '" + os.path.expandvars("$HOME/desenv/nocopyright/data/AJUDA_en.mkv") + "'\n");
-                        path_buffer = "/tmp/video_legenda_" + language["language"] + "_" + legenda.id + ".mkv";
+                    for i in range(len(self.legendas)):
+                        print(str(len(self.legendas) - 1), "/", i);
+                        #self.__make_video__(self.legendas[i], language);
+                        self.legendas[i].make_video( self.path_video, language, self.project.effects ); 
+                        path_buffer = "/tmp/video_legenda_" + language["language"] + "_" + self.legendas[i].id + ".mkv";
                         if os.path.exists(path_buffer):
                             f.write("file '" + path_buffer + "'\n");
+                            if i < len(self.legendas) - 1:
+                                diferenca_segundos = int(self.legendas[i + 1].inicio_milisegundo/1000) - int(self.legendas[i].fim_milisegundo/1000 );
+                                print("diferença: \033[91m", self.legendas[i + 1].inicio_milisegundo/1000 - self.legendas[i].fim_milisegundo/1000, "\033[0m" );
+                                if diferenca_segundos > 0:
+                                    path_out_silencio = "/tmp/silencio_" + str(uuid.uuid4()) + ".mkv";
+                                    insert_blank_audio(self.path_video, path_out_silencio, int(self.legendas[i].fim_milisegundo/1000 ), int(self.legendas[i + 1].inicio_milisegundo/1000))
+                                    f.write("file '" + path_out_silencio + "'\n");
+                        else:
+                            print("\033[91m", "FALHA NAO FOI ENCONTRADO UM VIDEO", self.legendas[i].texto, "\033[0m"); #  
+
                 command = "ffmpeg -f concat -safe 0 -i '/tmp/"+ self.id +".txt' -c copy '/tmp/final_"+ language["language"] +"_"+ self.id +".mkv' > /dev/null 2>&1"; #
                 os.system( command );
                 dir_final = os.path.join( self.directory, language["directory"] ); 
@@ -229,7 +195,6 @@ class Video:
                 os.system("ffmpeg -i '"+ path_mkv +"' -c:v libx264 -pix_fmt yuv420p -c:a copy '"+ path_mp4 + "'  > /dev/null 2>&1")
                 #os.system("ffmpeg -i '"+ path_mkv +"' '"+ path_mp4 + "' > /dev/null 2>&1" ) ;
             if os.path.exists( path_mkv ) and os.path.exists( path_mp4 ):
-                print("REMOVENDO: ", path_mkv);
                 os.unlink( path_mkv );
 
     @staticmethod
@@ -250,7 +215,6 @@ class Legenda:
     def __init__(self, texto, inicio, fim, fps, index=0):
         self.id = str( uuid.uuid4()  ) + str(index);
         self.texto = texto.strip();
-        #self.traduzido = None;
         self.inicio = inicio;
         self.fps = fps;
         self.fim = fim;
@@ -278,6 +242,7 @@ class Legenda:
             insert_audio_in_video(path_video, path_mp3, path_out, int(self.inicio_milisegundo/1000), int( self.fim_milisegundo/1000 ) )
             os.unlink(path_mp3);
             return path_out;
+        print("\033[91m", "FALHA NAO FOI ENCONTRADO UM AUDIO", "\033[0m");
         return None;
 
     def make_audio(self, translated_text, to_language, engine_tts, effects=[]):
@@ -290,14 +255,14 @@ class Legenda:
             gravar_kokoro( translated_text, to_language, path_mp3);
         else:
             return None;
-        for effect in effects:
-            path_mp3_buffer = "/tmp/effect_" + to_language["language"] + "_" + self.id + ".mp3";
-            command = "ffmpeg -i '"+ path_mp3 +"' -af 'rubberband=tempo=1.0:pitch="+ effect["value"] +":pitchq=quality' '" + path_mp3_buffer + "' > /dev/null 2>&1";
-            #print("\033[97m", command, "\033[0m");
-            os.system( command );
-            if os.path.exists(path_mp3):
-                os.unlink(path_mp3);
-            shutil.move(path_mp3_buffer, path_mp3);
+        if to_language.get("effects") != None:
+            for effect in to_language["effects"]:
+                path_mp3_buffer = "/tmp/effect_" + to_language["language"] + "_" + self.id + ".mp3";
+                command = "ffmpeg -i '"+ path_mp3 +"' -af 'rubberband=tempo=1.0:pitch="+ effect["value"] +":pitchq=quality' '" + path_mp3_buffer + "' > /dev/null 2>&1";
+                os.system( command );
+                if os.path.exists(path_mp3):
+                    os.unlink(path_mp3);
+                shutil.move(path_mp3_buffer, path_mp3);
         if os.path.exists( path_mp3 ):
             return path_mp3;
         return None;
@@ -306,7 +271,6 @@ class Legenda:
         try:
             from_code = from_language;
             to_code = to_language["language"];
-            print(from_code, "->", to_code, ": ");
             if to_code == from_code:
                 self.traducoes[to_language["language"]] = self.texto;
             else:
@@ -343,7 +307,6 @@ class Legenda:
     def load(path_file, fps):
         #teste = 0;
         filename, file_extension = os.path.splitext( path_file );
-        #print("\033[96m", path_file,"\033[0m" );
         legendas = [];
         linhas = open(path_file, "r").readlines();
         if file_extension == ".vtt":
@@ -374,8 +337,6 @@ class Legenda:
                 if os.path.exists(path_file):
                     os.unlink(path_file);
                 return None;
-        #for legenda in legendas:
-        #    print( legenda.to_string() );
         return legendas;
 
 # ----------------- grande gambibarra --------------------- CANSADO PRA CARALEO
@@ -389,7 +350,6 @@ def transcrever(path_video):
     model_size = "large"
     model = WhisperModel(model_size, device="cpu", compute_type="int8")
     segments, info = model.transcribe(path_video, beam_size=5)
-    #print("Detected language '%s' with probability %f" % (info.language, info.language_probability))
     with open(path_vtt, "w") as f:
         for segment in segments:
             end_datet_time = dt.now();
@@ -400,7 +360,6 @@ def transcrever(path_video):
             start = segment.start; #= str(segment.start).replace(".", ":") + ".0";
             end   = segment.end; # str(segment.end  ).replace(".", ":") + ".0";
             text = segment.text.strip();
-            #print(start, "-->", end, " :", segment.text );
             f.write(f"{start:.2f} --> {end:.2f}\n")
             f.write(f"{text}\n\n")
     return path_vtt;
@@ -437,6 +396,7 @@ def gravar_kokoro(text, language, path_mp3):
     return path_mp3;
 
 def mp3_total_frames(filename, fps):
+    # se tem 1.9 segundo e 30 frames por segundo, dá para achar quantos frames precisa
     args=("ffprobe", "-show_entries", "format=duration","-i",filename)
     popen = subprocess.Popen(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     output, err = popen.communicate()
@@ -450,6 +410,40 @@ def mp3_total_frames(filename, fps):
     err = None;
     return int(total);
 
+def insert_blank_audio(path_video_base, path_out, second_start, second_end):
+    try:
+        video_base = cv2.VideoCapture( path_video_base );
+        fps = video_base.get(5);
+        frame_start = int(second_start  * fps);
+        frame_end   = int(second_end    * fps);
+        path_buffer_sem_audio   = "/tmp/" + str(uuid.uuid4()) + ".mkv";
+        path_audio_nulo         = "/home/uell/desenv/nocopyright/data/buffer.mp3";
+        path_buffer_com_audio   = path_out;
+        total_frames_video = frame_end - frame_start;
+        video_base.set(cv2.CAP_PROP_POS_FRAMES, frame_start - 1);  #https://stackoverflow.com/questions/33650974/opencv-python-read-specific-frame-using-videocapture
+        frame_width =  int(video_base.get(3));
+        frame_height = int(video_base.get(4));
+        frame_size =   (frame_width,frame_height);
+        fourcc = cv2.VideoWriter_fourcc('M','J','P','G');
+        video_output = cv2.VideoWriter(path_buffer_sem_audio, fourcc , fps, frame_size);
+        for i in range(total_frames_video):
+            success, image = video_base.read();
+            video_output.write( image );
+        video_output.release();
+        video_base.release();
+        video_output = None;
+        video_base = None;
+        fourcc = None;
+
+        command = "ffmpeg -i '"+ path_buffer_sem_audio +"' -i '"+ path_audio_nulo +"' -c copy -map 0:v:0 -map 1:a:0 '"+ path_buffer_com_audio +"' > /dev/null 2>&1"
+        os.system(command);
+        if os.path.exists(path_buffer_sem_audio):
+            os.unlink( path_buffer_sem_audio );
+        return path_buffer_com_audio;
+    except:
+        traceback.print_exc();
+    return False;
+
 def insert_audio_in_video(path_video_base, path_audio, path_out, second_start, second_end):
     try:
         video_base = cv2.VideoCapture( path_video_base );
@@ -457,7 +451,6 @@ def insert_audio_in_video(path_video_base, path_audio, path_out, second_start, s
         frame_start = int(second_start  * fps);
         frame_end   = int(second_end    * fps);
         path_buffer_sem_audio = "/tmp/" + str(uuid.uuid4()) + ".mkv";
-        #path_buffer_com_audio = "/tmp/COM_audio" + str(uuid.uuid4()) + ".mkv";
         path_buffer_com_audio = path_out;
         total_frames_audio = mp3_total_frames( path_audio, fps ); #TODO arquivo nao existe...
         total_frames_video = frame_end - frame_start;
@@ -469,12 +462,11 @@ def insert_audio_in_video(path_video_base, path_audio, path_out, second_start, s
             quadro_mod = total_frames_video;
         
         video_base.set(cv2.CAP_PROP_POS_FRAMES, frame_start - 1);  #https://stackoverflow.com/questions/33650974/opencv-python-read-specific-frame-using-videocapture
-        #if video_output == None:
         frame_width = int(video_base.get(3));
         frame_height = int(video_base.get(4));
         frame_size = (frame_width,frame_height);
-        fourcc = cv2.VideoWriter_fourcc(*"MP4V")
-        fourcc = cv2.VideoWriter_fourcc(*'X264');
+        #fourcc = cv2.VideoWriter_fourcc(*"MP4V")
+        #fourcc = cv2.VideoWriter_fourcc(*'X264');
         fourcc = cv2.VideoWriter_fourcc('M','J','P','G');
         video_output = cv2.VideoWriter(path_buffer_sem_audio, fourcc , fps, frame_size);
         adicionados = 0;
@@ -491,7 +483,7 @@ def insert_audio_in_video(path_video_base, path_audio, path_out, second_start, s
                         video_output.write( image );
                         adicionados = adicionados + 1;
                 else:
-                    if i % quadro_mod != 0:
+                    if i % quadro_mod == 0:
                         video_output.write( image );
                         adicionados = adicionados + 1;
         video_output.release();
@@ -499,8 +491,7 @@ def insert_audio_in_video(path_video_base, path_audio, path_out, second_start, s
         video_output = None;
         video_base = None;
         fourcc = None;
-        #command = "ffmpeg -i '"+ path_buffer_sem_audio +"' -i '"+ path_audio +"' -c copy -map 0:v:0 -map 1:a:0 '"+ path_buffer_com_audio +"'  > /dev/null 2>&1"
-        command = "ffmpeg -i '"+ path_buffer_sem_audio +"' -i '"+ path_audio +"' -c:v copy -c:a aac -map 0:v:0 -map 1:a:0 '"+ path_buffer_com_audio +"'  > /dev/null 2>&1" 
+        command = "ffmpeg -i '"+ path_buffer_sem_audio +"' -i '"+ path_audio +"' -c:v copy -c:a aac -map 0:v:0 -map 1:a:0 '"+ path_buffer_com_audio +"' > /dev/null 2>&1 " 
         os.system(command);
         if os.path.exists(path_buffer_sem_audio):
             os.unlink( path_buffer_sem_audio );
